+++ CONTRIBUTING.md
@@ -7,9 +7,9 @@ These steps will guide you through contributing to this project:
 - Fork the repo
 - Clone it and install dependencies
 
-		git clone https://github.com/skogsmaskin/passport-zeit
+		git clone https://github.com/skogsmaskin/passport-vercel
 		npm install
 
 Make and commit your changes. Make sure the commands npm run build and npm run test:prod are working.
 
-Finally send a [GitHub Pull Request](https://github.com/skogsmaskin/passport-zeit/compare?expand=1) with a clear list of what you've done (read more [about pull requests](https://help.github.com/articles/about-pull-requests/)). Make sure all of your commits are atomic (one feature per commit).
+Finally send a [GitHub Pull Request](https://github.com/skogsmaskin/passport-vercel/compare?expand=1) with a clear list of what you've done (read more [about pull requests](https://help.github.com/articles/about-pull-requests/)). Make sure all of your commits are atomic (one feature per commit).
