+++ README.md
@@ -1,11 +1,11 @@
-# passport-zeit
+# passport-vercel
 [Passport](http://passportjs.org/) strategy for authenticating with
-[ZEIT](https://www.zeit.co/) using the OAuth 2.0 API.
+[Vercel](https://vercel.com/) using the OAuth 2.0 API.
  
-[![Build Status](https://travis-ci.com/skogsmaskin/passport-zeit.svg?branch=master)](https://travis-ci.com/skogsmaskin/passport-zeit)
+[![Build Status](https://travis-ci.com/skogsmaskin/passport-vercel.svg?branch=master)](https://travis-ci.com/skogsmaskin/passport-vercel)
 
-This module lets you authenticate with ZEIT in your Node.js applications.
-By plugging into Passport, ZEIT authentication can be easily and
+This module lets you authenticate with [Vercel](https://vercel.com/) in your Node.js applications.
+By plugging into Passport, Vercel authentication can be easily and
 unobtrusively integrated into any application or framework that supports
 [Connect](http://www.senchalabs.org/connect/)-style middleware, including
 [Express](http://expressjs.com/).
@@ -14,19 +14,19 @@ unobtrusively integrated into any application or framework that supports
 ## Install
 
 ```bash
-$ npm install passport-zeit
+$ npm install passport-vercel
 ```
 
 ## Usage
 
 #### Create an Application
 
-Before using `passport-zeit`, you must register an application with ZEIT.
-At the time being this is a manual process where you will have to contact ZEIT and get them to do it for you. Your application will be issued a client ID and client secret, which need to be provided to the strategy. You will also need to configure a callback URL which matches the route in your application.
+Before using `passport-vercel`, you must register an application with Vercel.
+At the time being this is a manual process where you will have to contact Vercel and get them to do it for you. Your application will be issued a client ID and client secret, which need to be provided to the strategy. You will also need to configure a callback URL which matches the route in your application.
 
 #### Configure Strategy
 
-The ZEIT authentication strategy authenticates users using a ZEIT account
+The Vercel authentication strategy authenticates users using a Vercel account
 and OAuth 2.0 tokens. The client ID and secret obtained when creating an
 application are supplied as options when creating the strategy. The strategy
 also requires a `verify` callback, which receives the access token, as well as 
@@ -34,23 +34,23 @@ also requires a `verify` callback, which receives the access token, as well as
 The `verify` callback must call `cb` providing a user to complete authentication.
 
 ```js
-const ZeitStrategy = require('passport-zeit').Strategy;
+const VercelStrategy = require('passport-vercel').Strategy;
 
-passport.use(new ZeitStrategy({
-    clientID: ZEIT_CLIENT_ID,
-    clientSecret: ZEIT_CLIENT_SECRET,
-    callbackURL: "http://127.0.0.1:3000/auth/zeit/callback",
+passport.use(new VercelStrategy({
+    clientID: VERCEL_CLIENT_ID,
+    clientSecret: VERCEL_CLIENT_SECRET,
+    callbackURL: "http://127.0.0.1:3000/auth/vercel/callback",
     state: true // use OAuth2 state param to protect against csrf attacks (requries express-session)
   },
   (accessToken, refreshToken, profile, cb) => {
-    User.findOrCreate({provider: 'zeit': providerId: profile.id }, (err, user) => {
+    User.findOrCreate({provider: 'vercel', providerId: profile.id }, (err, user) => {
       return cb(err, user)
     })
   }
 ))
 ```
 
-*Note: While ZEIT doesn't support refresh tokens (as for Jan. 2019),
+*Note: While Vercel doesn't support refresh tokens (as for Jan. 2019),
 the verify callback still uses a second parameter `refreshToken` to conform
 to the OAuth2 standard. This makes it easier to  share a verify callback function
 for all `passport-oauth2` based authentications you use with Passport.js.
@@ -58,17 +58,17 @@ The `refreshToken` will be undefined with this strategy, and should just be igno
 
 #### Authenticate Requests
 
-Use `passport.authenticate()`, specifying the `'zeit'` strategy, to
+Use `passport.authenticate()`, specifying the `'vercel'` strategy, to
 authenticate requests.
 
 For example, as route middleware in an [Express](http://expressjs.com/)
 application:
 
 ```js
-app.get('/auth/zeit', passport.authenticate('zeit'))
+app.get('/auth/vercel', passport.authenticate('vercel'))
 
-app.get('/auth/zeit/callback', 
-  passport.authenticate('zeit', { failureRedirect: '/login' }),
+app.get('/auth/vercel/callback', 
+  passport.authenticate('vercel', { failureRedirect: '/login' }),
   (req, res) => {
     // Successful authentication, redirect home.
     res.redirect('/');
